Avoided recurring letter addition to blacklist
 Say the actual word has a letter 'R' only once and guess has 'R' twice in two different positions

Say the position of one of these 'R's in the guess is right and other one is wrong. Now the letter 'R' in the wrong position gets added to the blacklist as wordle marks it as absent.

This in turn skips all words containing 'R' thereby making the algorithm run out of guesses

Avoided the above by filtering the blacklist off letters present at the right position after every evaluation